backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
openpyxl
requests
pandas
python-multipart
